Initial Commit
Initial commit

Jupyter notebook work documenting the progress of the project.

At some point in the project --- Will create a separate Notebook which streamlines the computations because the way the current book is written is for exploratory + documentative purpose (ie: I describe in markdown the purpose of blocks and discoveries and things I plan to do).