V1
Python code complete and dashboard created. Next steps would be to clean up the python code to be more presentable and directly coding the necessary elements. Current state of the notebook documents all the pitfalls and roadblocks that had to be dealt with.

Uploading image of finaldashboard to embed into readme file.